+++ src/main/java/com/szymon/apka/service/TeacherService.java
@@ -1,7 +1,8 @@
 package com.szymon.apka.service;
 
-import com.szymon.apka.entity.Student;
 import com.szymon.apka.entity.Teacher;
+import com.szymon.apka.exception.RelationStudentTeacherExistsException;
+import com.szymon.apka.repository.StudentTeacherRelationRepository;
 import com.szymon.apka.repository.TeacherRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -14,8 +15,11 @@ public class TeacherService {
 
     private final TeacherRepository teacherRepository;
 
-    public TeacherService(TeacherRepository teacherRepository) {
+    private final StudentTeacherRelationRepository relationRepository;
+
+    public TeacherService(TeacherRepository teacherRepository, StudentTeacherRelationRepository relationRepository) {
         this.teacherRepository = teacherRepository;
+        this.relationRepository = relationRepository;
     }
 
     public Teacher addTeacher(Teacher teacher) {
@@ -24,8 +28,14 @@ public class TeacherService {
 
     public boolean deleteTeacher(Long id) {
 
+        if(this.relationRepository.checkIfRelationExistsForTeacherId(id) == 1) {
+            throw new RelationStudentTeacherExistsException("Can't delete Teacher with id: " + id + " , because relation exists");
+        }
+
         if (this.teacherRepository.existsById(id)) {
             this.teacherRepository.deleteById(id);
+
+            return true;
         }
 
         return false;
